+++ pkgIndex.tcl
@@ -1,5 +1,5 @@
 
-package ifneeded apave 3.2.8 [list source [file join $dir apaveinput.tcl]]
+package ifneeded apave 3.3a3 [list source [file join $dir apaveinput.tcl]]
 
 # A short intro (for Ruff! docs generator:)